flow_manager should work. updated headers. alg should work. Progress. The indomitable will of mankind.

+++ flow_manager.h
@@ -1,19 +1,20 @@
 #ifndef TOKYOFLOWMANAGER
 #define TOKYOFLOWMANAGER
 
-
-
+#include <set>
 
 class flow_manager_t
 {
    private:
-   	private static std::set<uint32_t> used_udp_srcprts;
+   	private static std::set<uint16_t> used_udp_srcprts;
 
    public:
+
    	flow_manager_t(){
 
    	}
-   	uint32_t get_new_srcport();
+
+   	uint16_t get_new_srcport();
 
 };
